Add real name properties and set sane owner/group/mode defaults
Instead of passing nil for owner/group/mode set defaults in the resource. This makes it a lot more clear in any failure messaging what we're trying to do and in general it seems like the right way to do it. This also removes the old :name named name properties (say that 10 times fast) and replaces them with property name properties for overriding the resource name. This is why we have this Foodcritic rule. It alerts us to code smell.

Signed-off-by: Tim Smith <tsmith@chef.io>